+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/OnDeathTrigger.cs
@@ -12,7 +12,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers
     /// <summary>Represents an On Death trigger.</summary>
     public class OnDeathTrigger : Trigger, IHasTargetGroupID
     {
-        public override int ObjectID => (int)Enumerations.GeometryDash.TriggerType.OnDeath;
+        public override short ObjectID => (short)(int)Enumerations.GeometryDash.TriggerType.OnDeath;
         
         /// <summary>The target Group ID of the trigger.</summary>
         public short TargetGroupID { get; set; }
