+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/SpawnTrigger.cs
@@ -12,7 +12,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers
     /// <summary>Represents a Spawn trigger.</summary>
     public class SpawnTrigger : Trigger, IHasTargetGroupID
     {
-        public override int ObjectID => (int)Enumerations.GeometryDash.TriggerType.Spawn;
+        public override short ObjectID => (short)(int)Enumerations.GeometryDash.TriggerType.Spawn;
         
         /// <summary>The target Group ID of the trigger.</summary>
         public short TargetGroupID { get; set; }
